"Ozbourne Tower made into Tower child"
